New build system
I moved a lot of files around and completely replaced the existing
makefile.  It is no longer recursive.  The makefile in the root
directory knows how to build everything.

**Make Rules**
arm - *Default* Just builds code needed to run on ARM targets.
host - Builds core lib and unit tests which run on host build machine.
all - Combines arm and host.
gcov - Builds core libs and unit tests with code coverage enabled.
clean - Cleans out all output files.

I also added a deploy script to archive the new script that I now use
for building and deploying MRI to my GCC4MBED directory.

+++ tests/mocks/platformMock.cpp
@@ -78,7 +78,7 @@ void Platform_Init(Token* pParameterTokens)
 // Platform_Comm* Instrumentation
 static const char  g_emptyPacket[] = "$#00";
 static Buffer      g_receiveBuffers[2];
-static int         g_receiveIndex;
+static size_t      g_receiveIndex;
 static char*       g_pAlloc1;
 static char*       g_pAlloc2;
 static char*       g_pTransmitDataBufferStart;
